Do not cast accelerator command to float
Prior to this commit, the accelerator command was coming in as a
uint16_t, being cast to a float, being calculated upon, and then being
cast back to a uint16_t when controlling the vehicle. This commit keeps
the value as a uint16_t for clarity's sake.

+++ platforms/kia_soul/firmware/throttle/include/throttle_control.h
@@ -108,8 +108,8 @@ typedef struct
     bool operator_override; /* Flag indicating whether accelerator was manually
                                pressed by operator. */
 
-    float commanded_accelerator_position; /* Position of accelerator commanded
-                                             by the controller. */
+    uint16_t commanded_accelerator_position; /* Position of accelerator commanded
+                                                by the controller. */
 } kia_soul_throttle_control_state_s;
